Refactor: common laskenta-state-button and move some valintalaskenta translations
