+++ src/app/lokalisaatio/fi.json
@@ -513,13 +513,7 @@
     "valinta-save-error": "Valinnan tietojen tallentaminen epäonnistui",
     "valinta-save-success": "Valinnan tietojen tallentaminen onnistui",
     "pistesyotto": "Pistesyöttö",
-    "suorita-valintalaskenta": "Suorita valintalaskenta",
-    "keskeyta-valintalaskenta": "Keskeytä valintalaskenta",
     "valintalaskenta": "Valintalaskenta",
-    "sulje-laskennan-tiedot": "Sulje laskennan tiedot",
-    "piilota-suorittamattomat-hakukohteet": "Piilota suorittamattomat hakukohteet",
-    "nayta-suorittamattomat-hakukohteet": "Näytä suorittamattomat hakukohteet",
-    "syy": "Syy",
     "valintalaskenta-epaonnistui": "Valintalaskenta epäonnistui",
     "hakukohteita-valmiina": "Hakukohteita valmiina",
     "suorittamattomia-hakukohteita": "Suorittamattomia hakukohteita",
@@ -550,6 +544,12 @@
     "organizationLatestByTag": "Hakukohteen viimeisin"
   },
   "valintalaskenta": {
+    "suorita-valintalaskenta": "Suorita valintalaskenta",
+    "keskeyta-valintalaskenta": "Keskeytä valintalaskenta",
+    "piilota-suorittamattomat-hakukohteet": "Piilota suorittamattomat hakukohteet",
+    "nayta-suorittamattomat-hakukohteet": "Näytä suorittamattomat hakukohteet",
+    "sulje-laskennan-tiedot": "Sulje laskennan tiedot",
+    "syy": "Syy",
     "hyvaksymiskirjeet": {
       "eihakijoita": "Ei ole hakijoita, joille kirjeet voisi muodostaa"
     },
